Print argument to F_GETLK/SETLK/SETLKW as a pointer, not int.

+++ lib/libsysdecode/sysdecode.h
@@ -76,7 +76,7 @@ void	sysdecode_capfcntlrights(FILE *_fp, uint32_t _rights);
 void	sysdecode_capname(FILE *_fp, cap_rights_t *_rightsp);
 void	sysdecode_extattrnamespace(FILE *_fp, int _namespace);
 void	sysdecode_fadvice(FILE *_fp, int _advice);
-void	sysdecode_fcntl_arg(FILE *_fp, int _cmd, int _arg, int _base);
+void	sysdecode_fcntl_arg(FILE *_fp, int _cmd, uintptr_t _arg, int _base);
 bool	sysdecode_fcntl_arg_p(int _cmd);
 void	sysdecode_fcntl_cmd(FILE *_fp, int _cmd);
 void	sysdecode_fcntl_fileflags(FILE *_fp, int _flags);
